+++ cmd/docmgr/cmds/root.go
@@ -1,6 +1,7 @@
 package cmds
 
 import (
+	"github.com/go-go-golems/docmgr/pkg/completion"
 	"github.com/go-go-golems/docmgr/cmd/docmgr/cmds/changelog"
 	"github.com/go-go-golems/docmgr/cmd/docmgr/cmds/configcmd"
 	"github.com/go-go-golems/docmgr/cmd/docmgr/cmds/doc"
@@ -34,6 +35,9 @@ Helpful docs (built-in):
 	}
 
 	help_cmd.SetupCobraRootCommand(helpSystem, rootCmd)
+
+	// Enable carapace dynamic completion (adds hidden `_carapace` and bridges cobra)
+	completion.Attach(rootCmd)
 
 	if err := workspace.Attach(rootCmd); err != nil {
 		return nil, err